Add also indirect dependencies into module/include path
The "lxc" module depends on "security", which depends on "pam". Before
this commit, both "security" and "pam" had to be specified as deps in
lxc.yml. This is wrong, because the fact that "security" depends on
"pam" is an internal thing of the "security" module and it shouldn't
leak elsewhere nor be described multiple times in different .yml files.

This commit fixes this problem in a general way by adding indirect
dependencies to module/include path too.