Remove the "No 'moves' section" warning
It turns out we have quite some YaST modules where we don't need to do
any restructuring but there are still YCP files to translate there. For
these modules, no "moves" section is present in their .yml file and a
false-positive warning was emitted. For me, this is enough to get rid of
such a warning.