Improve how file names are reported in error.log
Before this change, errors in the error log reported by the "ybc"
command contained only part of the filename of the affected file
relative to module's work directory. This made it hard to see which
module that file belonged to. On the other hand, errors reported by the
"ruby" command contained full path to the affected file. These were
sometimes too long.

This commit unifies the behavior and writes the filename into the error
log in format like this:

  [autoinstallation] src/clients/inst_autosetup_upgrade.ycp

The module name is obvious, the path is always relative to module's work
directory, and the line is easily greppable.