Don't create packages in OBS automatically
Any error when running "osc co" in the "package" command context was
assumed to mean that a package doesn't exist in OBS and it needs to be
created. This was unfortunate because it masked some errors. For
example, if OSC wasn't configured on user's machine, "osc co" failed and
the user was then asked for a password (presumably by "osc meta"). The
real error message wasn't ever displayed.

Instead of trying to make the error handling more specific, I decided to
jsut throw the whole error handler and package creation part away. It
was bit hacky anyway and we already created most (if not all) OBS
packages we needed.