build order: fix network (hack within the flawed design)
The problem is exposed in yast2-network if Dir returns ISDN.ycp
earlier than NetHwDetection.ycp. It turns out that sorting the
directory listings hides the problem in all modules so we do just
that.

The problem is that when we attempt to break the module-include
dependency cycles we lose a module-include-module dependency. A way
to fix it would be to recurse over the includes first and solve the
dependencies over modules only (like the original ycpmakedep does!)
but the "sort" hack is a better effort-quality tradeoff.