Allow compiling of non-standalone include files using wrappers
Some include files are not standalone and need to be compiled in context
of another file (called a *wrapper*). This commit allows specifying
these include files and their wrappers in the "include_wrappers" section
in module metadata. YCP Killer then uses this information to execute Y2R
on the wrapper with a special flag (--extract-file), which makes it to
generate Ruby code only for statements from the include file.