Translate include files in testsuite correctly
Include files in the testsuite (which always have the .yh extension)
aren't in any exported directory, so they weren't considered as include
files by our code until now.

This commit extends the logic so that it always considers *.yh files as
include files, wherever they are. It also makes sure that sane
--extract-file option value is passed to Y2R for them.