Add --ref, --branch and --tag options to "yk clone"
The --branch and --tag options are equal to --ref, they are just
provided for readability (this is inspired by Bundler).

These options will allow YCP Killer to be used even after the final
conversion to Ruby was done in master. This might be useful e.g. for
porting SLE fixes (which must be done in YCP) to master.